+++ package.json
@@ -25,7 +25,7 @@
     "@angular/platform-browser-dynamic": "~10.0.4",
     "@angular/router": "~10.0.4",
     "@ng-bootstrap/ng-bootstrap": "^7.0.0",
-    "@ng-select/ng-select": "^4.0.4",
+    "@ng-select/ng-select": "^5.0.1",
     "@worldskills/bootstrap": "^4.4.1",
     "angular-oauth2-oidc": "^10.0.3",
     "barrelsby": "^2.2.0",
@@ -33,6 +33,7 @@
     "faker": "^4.1.0",
     "font-awesome": "^4.7.0",
     "jasmine": "^3.5.0",
+    "jquery": "^3.5.1",
     "ng-select": "^1.0.2",
     "rxjs": "~6.5.5",
     "terser": "^4.8.0",
