+++ projects/worldskills-angular-lib/assets/styles/index.css
@@ -1,0 +1,1 @@
+@import "form.css";
