+++ projects/worldskills-angular-lib/src/lib/date/datetime-picker/datetime-picker.component.html
@@ -1,6 +1,6 @@
 <div class="input-group mr-2">
   <input class="form-control form-control-sm" (blur)="inputBlur($event)"
-    [ngModel]="value.toString() | date:inputDatetimeFormat" (change)="onInputChange($event)" [disabled]='disabled'
+    [ngModel]="value ? value.toString() : '' | date:inputDatetimeFormat" (change)="onInputChange($event)" [disabled]='disabled'
     [placeholder]="inputDatetimeFormat" />
 
   <div class="input-group-append">
