+++ projects/worldskills-angular-lib/src/lib/entity-tree-select/components/entity-tree-select.component.spec.ts
