+++ projects/worldskills-angular-lib/src/lib/entity-tree-select/consts/default-labels.ts
