+++ projects/worldskills-angular-lib/src/lib/entity-tree-select/models/entity-tree-select-labels.ts
