+++ projects/worldskills-angular-lib/src/lib/entity-tree-select/types/tree-entity.ts
