+++ projects/worldskills-angular-lib/src/lib/select/components/select-async-header.component.ts
@@ -1,0 +1,16 @@
+import {Component, Input, TemplateRef, ViewChild} from '@angular/core';
+
+@Component({
+    selector: 'ws-select-async-header',
+    template: '<ng-template #templateRef><ng-container *ngIf="isString">{{content}}</ng-container><ng-container *ngIf="!isString"><ng-template [ngTemplateOutlet]="content"></ng-template></ng-container></ng-template>',
+})
+export class SelectAsyncHeaderComponent {
+
+    @Input() content: string | TemplateRef<any> = '';
+    @ViewChild('templateRef', {static: true}) templateRef;
+
+    get isString(): boolean {
+        return typeof this.content === 'string';
+    }
+
+}
