+++ projects/worldskills-angular-lib/src/lib/select/models/select-async-search-observer-resolver.ts
@@ -1,0 +1,6 @@
+import {Observable, PartialObserver} from 'rxjs';
+
+export type SelectAsyncSearchObserverResolver =
+    PartialObserver<string>
+    | ((value: string) => void)
+    | ((value: string) => Observable<string>);
