+++ projects/worldskills-angular-lib/src/lib/select/models/select-async-search-settings.ts
@@ -1,0 +1,4 @@
+import {SelectAsyncSearchObserverResolver} from './select-async-search-observer-resolver';
+import {SelectAsyncSearchSettingsAdvanced} from './select-async-search-settings-advanced';
+
+export type SelectAsyncSearchSettings = SelectAsyncSearchSettingsAdvanced | SelectAsyncSearchObserverResolver;
